+++ code/module-08/m08-start/module08_loans.ts
@@ -1,0 +1,10 @@
+/*  TODO Create a the Loans namespace. */
+namespace Loans {
+  export interface Loan {
+    principle: number,
+    interestRate: number        //* Interest rate percentage (eg. 14 is 14%)
+  }
+  export interface ConventionalLoan extends Loan {
+    months: number      //* Total number of months
+  }
+}
