+++ code/module-08/m08-start/module08_main.ts
@@ -2,41 +2,43 @@
     Lab Start */
 
 /*  TODO Create a the Loans namespace. */
-namespace Loans {
-  export interface Loan {
-    principle: number,
-    interestRate: number        //* Interest rate percentage (eg. 14 is 14%)
-  }
-  export interface ConventionalLoan extends Loan {
-    months: number      //* Total number of months
-  }
-}
+// namespace Loans {
+//   export interface Loan {
+//     principle: number,
+//     interestRate: number        //* Interest rate percentage (eg. 14 is 14%)
+//   }
+//   export interface ConventionalLoan extends Loan {
+//     months: number      //* Total number of months
+//   }
+// }
 
-/*  TODO Create LoanPrograms namespace. */
-namespace LoanPrograms {
-  /*  TODO Update the calculateInterestOnlyLoanPayment function. */
-  export function calculateInterestOnlyLoanPayment(loanTerms: Loans.Loan): string {
-    let payment: number;
-    payment = loanTerms.principle * calculateInterestRate(loanTerms.interestRate);
-    return 'The interest only loan payment is ' + payment.toFixed(2);
-  }
+// /*  TODO Create LoanPrograms namespace. */
+// namespace LoanPrograms {
+//   /*  TODO Update the calculateInterestOnlyLoanPayment function. */
+//   export function calculateInterestOnlyLoanPayment(loanTerms: Loans.Loan): string {
+//     let payment: number;
+//     payment = loanTerms.principle * calculateInterestRate(loanTerms.interestRate);
+//     return 'The interest only loan payment is ' + payment.toFixed(2);
+//   }
 
-  /*  TODO Update the calculateConventionalLoanPayment function. */
-  export function calculateConventionalLoanPayment(loanTerms: Loans.ConventionalLoan): string {
-    let interest: number = calculateInterestRate(loanTerms.interestRate);
-    let payment: number;
-    payment = loanTerms.principle * interest / (1 - (Math.pow(1 / (1 + interest), loanTerms.months)));
-    return 'The conventional loan payment is ' + payment.toFixed(2);
-  }
+//   /*  TODO Update the calculateConventionalLoanPayment function. */
+//   export function calculateConventionalLoanPayment(loanTerms: Loans.ConventionalLoan): string {
+//     let interest: number = calculateInterestRate(loanTerms.interestRate);
+//     let payment: number;
+//     payment = loanTerms.principle * interest / (1 - (Math.pow(1 / (1 + interest), loanTerms.months)));
+//     return 'The conventional loan payment is ' + payment.toFixed(2);
+//   }
 
-  function calculateInterestRate(interestRate: number): number {
-    let interest: number = interestRate / 1200;
-    return interest
-  }
-}
+//   function calculateInterestRate(interestRate: number): number {
+//     let interest: number = interestRate / 1200;
+//     return interest
+//   }
+// }
 
 
 /*  TODO Add reference paths. */
+/// <reference path="module08_loans.ts" />
+/// <reference path="module08_loan-programs.ts"/>
 
 /*  TODO Update the function calls. */
